% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateMu.R
\name{EstimateMusSaturated_PolC_from_MMR}
\alias{EstimateMusSaturated_PolC_from_MMR}
\title{Title}
\usage{
EstimateMusSaturated_PolC_from_MMR(mut_acc_experiment_data)
}
\arguments{
\item{mut_acc_experiment_data}{}
}
\description{
Title
}
