% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateMu.R
\name{EstimateMusSaturatedNopolC}
\alias{EstimateMusSaturatedNopolC}
\title{Title}
\usage{
EstimateMusSaturatedNopolC(mut_acc_experiment_data)
}
\arguments{
\item{mut_acc_experiment_data}{}
}
\description{
Title
}
