% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonWAIC.R
\name{PoissonWAIC_from_mcmc_samples_mu}
\alias{PoissonWAIC_from_mcmc_samples_mu}
\title{Title}
\usage{
PoissonWAIC_from_mcmc_samples_mu(
  mut_acc_data,
  mcmc_samples_mu,
  strain_order = c("WT3610", "polC", "MMR-", "polC_mutL")
)
}
\arguments{
\item{strain_order}{}
}
\description{
Title
}
