% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{SimulateDataDCM}
\alias{SimulateDataDCM}
\title{Simulate some data for one context}
\usage{
SimulateDataDCM(
  theta1 = 7.96451e-07,
  theta2 = 0.05348838,
  theta3 = 0.06330917,
  theta4 = 1.961348e-09,
  n = 21836,
  t_s = c(6453, 1895, 230, 248920),
  s = c("MMR-", "polC", "polC_mutL", "WT3610")
)
}
\arguments{
\item{s}{}
}
\description{
Simulate some data for one context
}
\examples{
theta = MomentMatchingEstimation(n = 21836, m_PolC_MMR = 4, t_PolC_MMR = 2.30e2, m_MMR = 6, t_MMR = 6.45e3, m_PolC = 2, t_PolC = 1.89e3, m_wt = 4, t_wt = 2.49e5)
SimulateData(theta1 = 7.96451e-07, theta2 = 0.05348838, theta3 = 6.330917e-02, theta4 = 1.961348e-09, n = 21836, t_s = c(6453, 1895, 230, 248920), s = c("MMR-", "polC", "polC_mutL", "WT3610"))
}
